agent/output/code/backend/requirements.txt
Flask>=2.3
python-dotenv>=1.0
mongoengine>=0.27
pymongo>=4.5
